Close SQL connection before module unload
This fixes reused connection error in MySQL when the driver module was not fully unloaded, whose event was the only responsible for closing it through the driver's destructor, and a new Connect was called from reused module instance upon level change. In SQLite3, the new effect is that no additional connection can be created for the same file, unlike previously.